fix: typings
